requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    make upgrade
#
asgiref==3.8.1
    # via
    #   -r requirements/base.txt
    #   django
bytecode==0.16.0
    # via ddtrace
cffi==1.17.1
    # via
    #   -r requirements/base.txt
    #   pynacl
click==8.1.7
    # via -r requirements/base.txt
coverage[toml]==7.6.9
    # via pytest-cov
ddt==1.7.2
    # via -r requirements/test.in
ddtrace==2.17.2
    # via -r requirements/test.in
deprecated==1.2.15
    # via opentelemetry-api
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   django-crum
    #   django-waffle
django-crum==0.7.9
    # via -r requirements/base.txt
django-waffle==4.2.0
    # via -r requirements/base.txt
envier==0.6.1
    # via ddtrace
factory-boy==3.3.1
    # via -r requirements/test.in
faker==33.1.0
    # via factory-boy
importlib-metadata==8.5.0
    # via opentelemetry-api
iniconfig==2.0.0
    # via pytest
mock==5.1.0
    # via -r requirements/test.in
newrelic==10.3.1
    # via
    #   -r requirements/base.txt
    #   -r requirements/test.in
opentelemetry-api==1.28.2
    # via
    #   -r requirements/test.in
    #   ddtrace
packaging==24.2
    # via pytest
pbr==6.1.0
    # via
    #   -r requirements/base.txt
    #   stevedore
pluggy==1.5.0
    # via pytest
protobuf==5.29.1
    # via ddtrace
psutil==6.1.0
    # via -r requirements/base.txt
pycparser==2.22
    # via
    #   -r requirements/base.txt
    #   cffi
pynacl==1.5.0
    # via -r requirements/base.txt
pytest==8.3.4
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==6.0.0
    # via -r requirements/test.in
pytest-django==4.9.0
    # via -r requirements/test.in
python-dateutil==2.9.0.post0
    # via faker
six==1.17.0
    # via python-dateutil
sqlparse==0.5.2
    # via
    #   -r requirements/base.txt
    #   django
stevedore==5.4.0
    # via -r requirements/base.txt
typing-extensions==4.12.2
    # via
    #   ddtrace
    #   faker
wrapt==1.17.0
    # via
    #   ddtrace
    #   deprecated
xmltodict==0.14.2
    # via ddtrace
zipp==3.21.0
    # via importlib-metadata
